Display generic tabular icon
Current thumbnail was using a tag lib to
re-render some tabular data from the
dataset. However, this gets in trouble
as actual dataset_embed.html template
leverages gn-map.js to replace gn-container

The tabular data is hacked in HTML so it
would be PITA to do this in a script tag.
The 'right' way to render tabular data
instead of a preview map would be to add
gn-tabular.js which does the preview similar
to gn-map.js and friends.

For now, it should be sufficient to rendere
an SVG as the actual data is previewed
as data content anyway.