Allows to display an array of anchor tags
DetailInfos component already is capable to render
an array of children for each DetailInfoField. For the
field type 'link' the DetailInfoField has been extended
to get rendering attributes from the actual field value
instead of the field (from config).

In case, a DetailInfoField of type 'link' misses the href
attribute from the localConfig, the actual rendering
attributes are expected to be provided from the context.